fix: move cli tool to seperate file
