+++ package.json
@@ -3,11 +3,15 @@
     "version": "0.4.0",
     "description": "Easily translate JSON into type definitions",
     "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "files": [
+        "dist"
+    ],
     "bin": {
-        "json2struct": "./dist/index.js"
+        "json2struct": "./dist/cli.js"
     },
     "scripts": {
-        "start": "node dist/index.js",
+        "start": "node dist/cli.js",
         "build": "tsc",
         "format": "prettier --ignore-path .gitignore --write .",
         "format:check": "prettier --check --ignore-path .gitignore --write .",
