+++ src/index.ts
@@ -1,15 +1,12 @@
-#!/usr/bin/env node
-
-import fs from 'fs/promises';
-
-import { Command, Option } from '@commander-js/extra-typings';
-
-import { generateJuliaStruct } from './languages/julia';
-import { generatePythonStruct } from './languages/python';
-import { generateTypeScriptType } from './languages/typescript';
+import { generateJuliaStruct, generatePythonStruct, generateTypeScriptType } from './languages';
 import { Token, tokenize } from './tokenizer';
 
-function convertToLanguage(language: string, token: Token) {
+export * from './languages';
+export * from './tokenizer';
+
+export type SupportedLanguage = 'typescript' | 'python' | 'julia';
+
+export function convertToLanguage(language: SupportedLanguage, token: Token) {
     switch (language) {
         case 'typescript':
             return generateTypeScriptType(token);
@@ -25,53 +22,15 @@ function convertToLanguage(language: string, token: Token) {
     }
 }
 
-const program = new Command();
+/**
+ *
+ * @param language the language to translate to
+ * @param json unparsed json string
+ */
+export default function json2struct(language: string, json: string) {
+    const parsedJson = JSON.parse(json);
 
-program
-    .name('json2struct')
-    .description('Easily translate JSON into type definitions')
-    .version('0.4.0')
-    .configureOutput({
-        writeOut: (str) => process.stdout.write(`[OUT] ${str}`),
-        writeErr: (str) => process.stdout.write(`[ERR] ${str}`),
-        outputError: (str, write) => write(`\x1b[31m${str}\x1b[0m`),
-    });
+    const tokens = tokenize(parsedJson);
 
-program
-    .command('convert <input>', { isDefault: true })
-    .description('Convert JSON file to type file')
-    .option('-o --output <output-file>')
-    .option('--overwrite')
-    .addOption(
-        new Option('-l --language <output-language>').choices(['typescript', 'python', 'julia']).default('typescript')
-    )
-    .action(async (inputPath, args) => {
-        console.info(`\u001b[32mjson2struct: Converting ${inputPath} to ${args.language}:\u001b[0m`);
-
-        if (!args?.output?.length && args?.overwrite) {
-            program.error('--overwrite options requires an output path');
-            return;
-        }
-
-        const fileContent = await fs.readFile(inputPath);
-
-        const json = JSON.parse(fileContent.toString());
-
-        const tokens = tokenize(json);
-
-        const generatedStruct = convertToLanguage(args?.language ?? 'typescript', tokens);
-
-        if (args.output?.length) {
-            if (args?.overwrite) {
-                await fs.writeFile(args.output, generatedStruct);
-            } else {
-                await fs.appendFile(args.output, generatedStruct);
-            }
-        }
-
-        console.info(generatedStruct);
-    });
-
-program.addHelpCommand();
-
-program.parse();
+    return convertToLanguage(language as SupportedLanguage, tokens);
+}
