+++ src/languages/index.ts
@@ -1,0 +1,3 @@
+export { generateTypeScriptType } from './typescript';
+export { generatePythonStruct } from './python';
+export { generateJuliaStruct } from './julia';
